exceptions handled for search and booking

+++ start_point/src/main/java/com/example/airline_api/services/FlightService.java
@@ -8,6 +8,7 @@ import com.example.airline_api.repositories.PassengerRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import javax.naming.directory.InvalidSearchFilterException;
 import java.util.List;
 import java.util.Optional;
 
@@ -33,20 +34,41 @@ public class FlightService {
         return flightRepository.findById(id);
     }
 
-    public Booking addPassengerToFlight(Long passengerId, Long flightId) {
-        Flight flight = flightRepository.getById(flightId);
-        Passenger passenger = passengerRepository.getById(passengerId);
+    public Booking bookPassengerOnFlight (Long passengerId, Long flightId) throws Exception{
+        Optional<Flight> optionalFlight = flightRepository.findById(flightId);
+        Optional<Passenger> optionalPassenger = passengerRepository.findById(passengerId);
+        if(optionalFlight.isEmpty()){
+            throw new IllegalArgumentException("Flight not found");
+        }
+        Flight flight = optionalFlight.get();
+        if(optionalPassenger.isEmpty()){
+            throw new IllegalArgumentException("Passenger not found");
+        }
+        Passenger passenger = optionalPassenger.get();
+        if (flight.getPassengers().contains(passenger)){
+            throw new Exception("Passenger already on flight");
+        }
+        if(flight.getPassengers().size()>=flight.getCapacity()){
+            throw new Exception("Flight full");
+        }
         flight.addPassenger(passenger);
         flightRepository.save(flight);
         return new Booking(passengerId,passenger.getName(),flightId,flight.getDestination());
     }
 
-
-    public void cancelBooking(Long id) {
+    public void cancelBooking(Long id) throws Exception{
+        if(flightRepository.findById(id).isEmpty()){
+            throw new Exception("No flight with this Id");
+        }
         flightRepository.deleteById(id);
     }
 
-    public List<Flight> searchFlight(String destination){
-        return flightRepository.findByDestinationIs(destination);
+    public List<Flight> searchFlight(String destination) throws InvalidSearchFilterException{
+        List<Flight> flights = flightRepository.findByDestinationIs(destination);
+        if (flights.isEmpty()){
+            throw new InvalidSearchFilterException("Flight not found");
+        }
+
+        return flights;
     }
 }
